fire_guard_project/requirements.txt
ultralytics
opencv-python
pyserial
